fix staging related scenarios

+++ app/controllers/application_controller.rb
@@ -236,7 +236,7 @@ class ApplicationController < ActionController::Base
   end
 
   def clear_old_data_from_staging
-    if ProductionEnvironment.staging?
+    if ProductionEnvironment.staging? && DatabaseHelper.postgres?
       ActiveRecord::Base.connection.execute('delete from user_sessions where updated_at < now()::date - 2')
     end
   end
